Introduce ppc64le support for OpenHPC
With this all spec files have been adapted to also build on ppc64le.
Right now only llvm-compilers.spec does not build on ppc64le with:

 llvm5-compilers-ohpc-5.0.0/llvm/projects/libunwind/include/__libunwind_config.h:55:4: error: #error "Unsupported architecture."

As this seems to be part of the actual source code and not packaging
related, and as nothing depends yet on llvm-compilers, llvm-compilers
has been ignored for the ppc64le effort.

This is also only compile tested but a first step towards ppc64le
support.

Test build results available at:

 https://copr.fedorainfracloud.org/coprs/adrian/ohpc-ppc64le/packages/

Signed-off-by: Adrian Reber <areber@redhat.com>